policy: respect FLAG_TURN_SCREEN_ON flag with device flip covers
If an application uses WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON
and the device supports covers and the cover is closed, then after the
screen would turn on due to this flag, it would immediately be turned off
because PhoneWindowManager thinks the cover is closed, so it puts the device to
sleep.

Respect this flag and do nothing if it is detected.

Change-Id: I4f1355f502e5bb41b0aba4ab9eaf98da99bfee0f
Signed-off-by: Kshitij Gupta <kshitijgm@gmail.com>